The skin files can read other files with @include statement.

+++ Library/TrayWindow.cpp
@@ -493,7 +493,7 @@ LRESULT CALLBACK CTrayWindow::WndProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARA
 				int pos = (wParam & 0x0ffff) - ID_THEME_FIRST;
 
 				const std::vector<std::wstring>& themes = Rainmeter->GetAllThemes();
-				if (pos >= 0 && pos < themes.size())
+				if (pos >= 0 && pos < (int)themes.size())
 				{
 					std::wstring command = L"\"" + Rainmeter->GetPath();
 					command += L"\\Addons\\RainThemes\\RainThemes.exe\" /load \"";
